Move remaining files from html to generator
