+++ lib/src/generator/resource_loader.dart
