+++ lib/src/generator/template_data.dart
